feat(job search): componentized job search
componentized job search and passed it into landing.jsx and search.jsx. Fixed Dice API so job type
is taken into consideration

BREAKING CHANGE: Not breaking, but location doesn't work still. Location for dice API takes a
zipcode for its location, and not city. So need to have a way to convert a cityname to a zipcode